refactor: change folder structure with calendarView
